Rename fp-bindgen-support feature
