+++ fp-bindgen-support/Cargo.toml
@@ -19,12 +19,12 @@ once_cell = "1"
 rmp-serde = "1.0.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_bytes = "0.11"
-wasmer = { version = "2.1", optional = true }
+wasmer = { version = "*", optional = true }
 thiserror = { version = "1.0.26", optional = true }
 tracing = "0.1.37"
 
 [features]
 default = []
 async = []
-host = ["wasmer", "thiserror"]
 guest = []
+wasmer2_host = ["dep:wasmer", "dep:thiserror"]
