+++ fp-bindgen-support/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod common;
 #[cfg(feature = "guest")]
 pub mod guest;
-#[cfg(feature = "host")]
-pub mod host;
 #[cfg(feature = "http")]
 pub mod http;
+#[cfg(feature = "wasmer2_host")]
+pub mod wasmer2_host;
 
 pub use fp_bindgen_macros::{fp_export_impl, fp_export_signature, fp_import_signature};
