+++ fp-bindgen-support/src/wasmer2_host/io.rs
